folders

+++ rpp/head.cpp
@@ -2,11 +2,11 @@
 
 namespace R {
 
-	variant_vector head(const variant_vector& x, int n) {
+	variant_vector head(const variant_vector& x, size_t n) {
 		return variant_vector (x.begin(), (x.size() < n) ? x.end() : x.begin() + 6);
 	}
 
-	data_frame head(const data_frame& x, int n) {
+	data_frame head(const data_frame& x, size_t n) {
 		data_frame df;
 		for (const auto& [key, vctr] : x) { // keys as column headings
 			df[key] = head(vctr, n);
