+++ rpp/tail.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "containers.h"
+#include "data_structures.h"
 
 namespace R {
 
@@ -8,11 +8,11 @@ namespace R {
 	/**
 	 * Returns the last n items of a variant vector
 	 */
-	variant_vector tail(const variant_vector& x, int n = 6);
+	variant_vector tail(const variant_vector& x, size_t n = 6);
 
 	/**
 	 * Returns the last n rows of a data frame
 	 */
-	data_frame tail(const data_frame& x, int n = 6);
+	data_frame tail(const data_frame& x, size_t n = 6);
 
 }
